move memory access logic out of ops
ops now just use their params as given. the execute() loop of the
Intcode Computer now determines what exactly to pass to any given opcode
based on the parameter modes in the instruction and the opcode's
writeParam. This feels like a good path, it let me delete a lot of wet
code from the opcodes. But the opcodes still feel like they know too
much about the computer. They might make more sense as members of
OpcodeComputer somehow, but I haven't thought of a way to invoke them by
their number that I like in that scenario. Either way, this feels like a
big improvement to the previous iteration of the Intcode Computer.